fix(docs): keep the server guard on the status fetch and stop caching failures
splitting the pure parser into lib/status-state.ts lets the test import it directly while lib/status.ts keeps its server-only marker, so a next revalidate fetch cannot be pulled into a client bundle later. traction-packages.test.ts works around the same tension by regex parsing its module as source text.

the route now returns 503 without cache headers when the state is unknown, matching app/api/npm/downloads/route.ts. serving null under s-maxage=300 with a 600s stale window pinned the fallback for up to fifteen minutes after a blip on the status host, which is exactly when the badge should be informative.
