+++ apps/docs/app/api/status/route.ts
@@ -1,9 +1,10 @@
 import { getStatusState } from "@/lib/status";
 
-export const revalidate = 300;
-
 export async function GET() {
   const state = await getStatusState();
+  if (state === null) {
+    return Response.json({ error: "status unavailable" }, { status: 503 });
+  }
   return Response.json(
     { state },
     {
