+++ apps/docs/lib/status-state.test.ts
@@ -1,4 +1,4 @@
-import { normalizeStatusState } from "./status";
+import { normalizeStatusState } from "./status-state";
 
 describe("normalizeStatusState", () => {
   it("maps the documented operational state", () => {
